feat(Status): Move usePTeamServiceDetailsData hook to Status directory for better organization
